fix: 우선순위 큐 연산 고려 점수 필드 타입 수정

+++ src/main/java/com/example/masterplanbbe/domain/recommendation/entity/Recommendation.java
@@ -6,8 +6,6 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Entity
 @Table(name = "recommendations")
 @Getter
@@ -25,8 +23,8 @@ public class Recommendation {
     private Spec spec;
 
     @Column(name = "old_score", nullable = false)
-    private BigDecimal oldScore;
+    private Double oldScore;
 
     @Column(name = "new_score", nullable = false)
-    private BigDecimal newScore;
+    private Double newScore;
 }
